Update stripejs.php
for hookOrderConfirmation, added 'order_pending' smarty variable so we could present a better/meaningful message to the customer when the charge succeeds, but verifications have failed.
for processPayment, we catch the Stripe CardError the correct way, and redirect back to payment selection page with a generic failure error.  Also fix the redirection URL in the existing Exception block, so the pages include .php.  This is required to properly support both PSv1.4 and v1.5
We also detect CVC failures, and create the order using the pending order status.