Update payment.php
Capture the paypal customer phone number if it exists. This is necessary since most stores require that a phone number is provided during checkout. Without this, Prestashop throws an exception when trying to create the Address object.

By default, Paypal does not send the phone number, so we must first do an 'isset' to determine if it was provided.

Also the Paypal merchant has to change their website preferences in Paypal profile to force Paypal to send it. Otherwise it will not be sent in the Paypal response