Update payment.php
2 changes required

First, Prestashop enforces that a phone number is required for all addresses, this is default functionality.  Therefore if Paypal returns a phone number, then we should save it to avoid exception being thrown

Second, not every Country has states, and therefore Paypal does not always return a 'SHIPTOSTATE'.  So before we attempt to assign the state, we should first check if the Country "containsStates"