PayPal Error 11812 Fix: Non Latin Characters
PayPal can handle up to 127 single byte characters for the description
of a product. Non Latin character languages like Greek or Hebrew have
characters that are double-byte. So the 100 character limit, in Greek
seems like a 200 single-byte character limit and that's the reason of
the Error 11812. This limits the description to 50 characters so that
the double-characters will be ok.