Possible security+usability issue in expresscheckout.php
Not sure if this is the best way to address it (a pull request) but here is the issue:

Right now the code will automatically log into the email provided by Paypal.  This is an issue because with Paypal you can have several emails associated with one account.  While testing with my store I logged in as original@site.com but the account after returning was new@site.com .  The order would not be associated with original@site.com as I was switched to new@site.com.  So not only does it not care about security, it's now no longer part of the account I originally wanted the order on and thus causing customer support nightmares.

The change proposed will check for a valid customer context and that an email exists (avoiding a blank customer object.)  The logging into an existing email without a password doesn't work.  From initial testing the change should be fine.  Open to suggestions though.